\name{find_min_vmode}
\alias{find_min_vmode}
\title{Find minimum vmode for non-ff x}
\usage{
  find_min_vmode(x, use.na = TRUE)
}
\arguments{
  \item{x}{any R type}

  \item{use.na}{require that the vmode handle NAs}
}
\value{
  int index of ff vmode that will hold x.
}
\description{
  Find minimum vmode for non-ff x
}

