\name{padNA}
\alias{padNA}
\title{Add a length of NAs before or after x
(used in Stacking)}
\usage{
  padNA(x, before = 0, after = 0)
}
\arguments{
  \item{x}{a data vector}

  \item{before}{length of NA to insert}

  \item{after}{length of NA to insert}
}
\value{
  same vector padded with NAs.
}
\description{
  Add a length of NAs before or after x (used in Stacking)
}

