\name{vecEqual}
\alias{vecEqual}
\title{Test equality and order of a vector}
\usage{
  vecEqual(x, y, order = FALSE)
}
\arguments{
  \item{x}{character}

  \item{y}{character}

  \item{order}{logical. stricter test for order of x and y}
}
\value{
  logical
}
\description{
  Test equality and order of a vector
}

