\name{vectorIntegrate}
\alias{vectorIntegrate}
\title{Integrate characters with careful ordering of result}
\usage{
  vectorIntegrate(x, y)
}
\arguments{
  \item{x}{character vector}

  \item{y}{character vector}
}
\value{
  union of x and y, with \emph{write something about
  expected return ordering}.
}
\description{
  Integrate characters with careful ordering of result
}

